+++ apps/docs/app/docs/[[...slug]]/page.tsx
@@ -13,9 +13,9 @@ import {CONTENT_PATH, TAG} from "@/libs/docs/config";
 import {getHeadings} from "@/libs/docs/utils";
 
 interface DocPageProps {
-  params: {
+  params: Promise<{
     slug: string[];
-  };
+  }>;
 }
 
 async function getDocFromParams({params}: DocPageProps) {
@@ -37,7 +37,8 @@ async function getDocFromParams({params}: DocPageProps) {
   return {doc, headings, currentRoute};
 }
 
-export async function generateMetadata({params}: DocPageProps): Promise<Metadata> {
+export async function generateMetadata(props: DocPageProps): Promise<Metadata> {
+  const params = await props.params;
   const {doc} = await getDocFromParams({params});
 
   if (!doc) {
@@ -77,7 +78,8 @@ export async function generateStaticParams(): Promise<DocPageProps["params"][]> 
   }));
 }
 
-export default async function DocPage({params}: DocPageProps) {
+export default async function DocPage(props: DocPageProps) {
+  const params = await props.params;
   const {doc, headings, currentRoute} = await getDocFromParams({params});
 
   if (!doc) {
