+++ apps/docs/content/blog/v2.6.0.mdx
@@ -2,7 +2,7 @@
 title: "NextUI v2.6.0 🔥"
 description: "NextUI v2.6.0 includes 4 new components, React 19 & Next.js 15 support, and lots of bug fixes and improvements."
 date: "2024-12-03"
-image: "/blog/v2.6.0.jpg"
+image: "/blog/v2.4.0.jpg"
 tags: ["nextui", "v2.6.0", "release", "bug fixes", "improvements", "animations"]
 author:
   name: "Junior Garcia"
@@ -14,7 +14,7 @@ author:
 import {formContent} from "@/content/components/form";
 
 <img
-  src="/blog/v2.6.0_2x.jpg"
+  src="/blog/v2.4.0_2x.jpg"
   width={700}
   height={350}
   alt="NextUI v2.6.0"
