+++ apps/docs/package.json
@@ -3,7 +3,7 @@
   "version": "2.0.0",
   "private": true,
   "scripts": {
-    "dev": "rimraf .contentlayer && concurrently \"contentlayer2 dev\" \"next dev\"",
+    "dev": "rimraf .contentlayer && concurrently \"contentlayer2 dev\" \"next dev --turbopack\"",
     "build": "contentlayer2 build && next build ",
     "build:analyze": "ANALYZE=true next build",
     "start": "next start",
@@ -18,8 +18,8 @@
   "dependencies": {
     "@codesandbox/sandpack-react": "^2.6.4",
     "@emotion/is-prop-valid": "^1.3.1",
-    "@iconify/icons-solar": "^1.2.3",
-    "@iconify/react": "^4.1.1",
+    "@iconify/icons-solar": "1.2.3",
+    "@iconify/react": "5.0.2",
     "@internationalized/date": "3.5.6",
     "@mapbox/rehype-prism": "^0.6.0",
     "@nextui-org/aria-utils": "workspace:*",
@@ -52,57 +52,59 @@
     "@react-stately/layout": "3.13.9",
     "@react-stately/tree": "3.8.1",
     "@rehooks/local-storage": "^2.4.5",
+    "@tanstack/react-virtual": "3.10.9",
     "canvas-confetti": "^1.9.2",
     "cmdk": "^0.2.0",
-    "color2k": "^2.0.2",
-    "contentlayer2": "^0.5.1",
-    "date-fns": "^2.30.0",
-    "framer-motion": "11.9.0",
-    "geist": "^1.3.1",
+    "color2k": "2.0.3",
+    "contentlayer2": "0.5.3",
+    "date-fns": "4.1.0",
+    "framer-motion": "11.11.13",
+    "geist": "1.3.1",
     "github-slugger": "^2.0.0",
     "gray-matter": "^4.0.3",
     "hast-util-to-html": "9.0.3",
     "marked": "^5.1.0",
-    "match-sorter": "^6.3.1",
-    "mini-svg-data-uri": "^1.4.3",
-    "mitt": "3.0.0",
-    "next": "13.5.1",
-    "next-contentlayer2": "^0.5.1",
-    "next-themes": "^0.2.1",
+    "match-sorter": "8.0.0",
+    "mini-svg-data-uri": "1.4.4",
+    "mitt": "3.0.1",
+    "next": "15.0.3",
+    "next-contentlayer2": "0.5.3",
+    "next-themes": "0.4.3",
     "nprogress": "^0.2.0",
     "parse-numeric-range": "1.2.0",
-    "posthog-js": "1.187.1",
+    "posthog-js": "1.184.1",
     "prism-react-renderer": "^1.2.1",
     "querystring": "^0.2.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-hook-form": "^7.51.3",
     "react-icons": "^4.10.1",
-    "react-live": "^2.3.0",
+    "react-live": "^4.1.8",
     "react-multi-ref": "^1.0.1",
     "react-wrap-balancer": "^1.0.0",
     "refractor": "3.3.1",
     "rehype": "13.0.2",
     "rehype-parse": "9.0.1",
     "rehype-pretty-code": "^0.14.0",
-    "rehype-slug": "^6.0.0",
+    "rehype-slug": "6.0.0",
     "remark": "^15.0.1",
-    "remark-gfm": "^4.0.0",
-    "rss": "^1.2.2",
-    "scroll-into-view-if-needed": "3.0.10",
+    "remark-gfm": "4.0.0",
+    "rss": "1.2.2",
+    "scroll-into-view-if-needed": "3.1.0",
     "sharp": "^0.32.1",
     "shelljs": "^0.8.4",
-    "swr": "^2.2.1",
+    "swr": "2.2.5",
     "tailwind-variants": "^0.1.20",
     "unified": "^11.0.5",
-    "unist-util-visit": "^5.0.0",
-    "usehooks-ts": "^2.9.1",
-    "zustand": "^4.3.8"
+    "unist-util-visit": "5.0.0",
+    "usehooks-ts": "3.1.0",
+    "zustand": "5.0.1"
   },
   "devDependencies": {
     "@docusaurus/utils": "2.0.0-beta.3",
-    "@next/bundle-analyzer": "^13.4.6",
-    "@next/env": "^13.4.12",
+    "@next/bundle-analyzer": "15.0.3",
+    "@next/env": "15.0.3",
+    "mdx-bundler": "10.0.3",
     "@react-types/calendar": "3.4.10",
     "@react-types/datepicker": "3.8.3",
     "@react-types/shared": "3.25.0",
@@ -116,22 +118,28 @@
     "@types/react": "18.2.8",
     "@types/react-dom": "18.2.4",
     "@types/refractor": "^3.0.2",
-    "@types/rss": "^0.0.30",
+    "@types/rss": "0.0.32",
     "@types/shelljs": "^0.8.9",
     "@types/uuid": "^8.3.1",
     "algoliasearch": "^4.10.3",
-    "autoprefixer": "^10.4.14",
+    "autoprefixer": "10.4.20",
     "dotenv": "^16.0.1",
     "esbuild-plugin-raw": "0.1.8",
-    "eslint-config-next": "^13.5.1",
+    "eslint-config-next": "15.0.3",
     "markdown-toc": "^1.2.0",
-    "next-sitemap": "^4.1.8",
+    "next-sitemap": "4.2.3",
     "node-fetch": "^3.2.10",
-    "postcss": "^8.4.21",
+    "postcss": "8.4.49",
     "prettier": "^2.7.1",
-    "tailwindcss": "^3.4.0",
+    "tailwindcss": "3.4.14",
     "tsx": "^3.8.2",
     "typescript": "^5.5.0",
     "uuid": "^8.3.2"
+  },
+  "pnpm": {
+    "overrides": {
+      "react": "18.2.0",
+      "react-dom": "18.2.0"
+    }
   }
 }
