+++ apps/docs/styles/globals.css
@@ -1,10 +1,10 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
 @import 'prism.css';
 @import 'nprogress.css';
 @import 'ads.css';
+
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
 
 html {
   font-size: 16px;
