+++ package.json
@@ -142,8 +142,8 @@
   },
   "pnpm": {
     "overrides": {
-      "react": "^18.2.0",
-      "react-dom": "^18.2.0",
+      "react": "18.2.0",
+      "react-dom": "18.2.0",
       "@changesets/get-github-info": "0.6.0"
     }
   },
